requirements.txt
tkinter
opencv-python
numpy
dtw-python
scipy
pygame
matplotlib
pandas
